+++ packages/dbx-web/src/lib/layout/content/content.border.directive.ts
@@ -1,4 +1,5 @@
-import { Directive } from '@angular/core';
+import { Directive, Input } from '@angular/core';
+import { DbxThemeColor } from '../style/style';
 
 /**
  * Section used to wrap content in a border with internal padding.
@@ -6,7 +7,11 @@ import { Directive } from '@angular/core';
 @Directive({
   selector: 'dbx-content-border,[dbxContentBorder]',
   host: {
-    class: 'd-block dbx-content-border'
+    class: 'd-block dbx-content-border',
+    '[class]': `"dbx-content-border-" + color`
   }
 })
-export class DbxContentBorderDirective {}
+export class DbxContentBorderDirective {
+  @Input()
+  color: DbxThemeColor = 'background';
+}