+++ packages/firebase/src/lib/common/firestore/snapshot/snapshot.field.ts
@@ -47,7 +47,9 @@ import {
   PrimativeKeyStringDencoderFunction,
   PrimativeKeyDencoderFunction,
   mapObjectMap,
-  UnitedStatesAddress
+  UnitedStatesAddress,
+  ZoomLevel,
+  DEFAULT_LAT_LNG_STRING_VALUE
 } from '@dereekb/util';
 import { FirestoreModelData, FIRESTORE_EMPTY_VALUE } from './snapshot.type';
 import { FirebaseAuthUserId } from '../../auth/auth';
@@ -640,8 +642,10 @@ export interface FirestoreLatLngStringConfig extends DefaultMapConfiguredFiresto
 
 /**
  * Default value used by firestoreLatLngString
+ *
+ * @Deprecated use DEFAULT_LAT_LNG_STRING_VALUE
  */
-export const DEFAULT_FIRESTORE_LAT_LNG_STRING_VALUE = '0,0';
+export const DEFAULT_FIRESTORE_LAT_LNG_STRING_VALUE = DEFAULT_LAT_LNG_STRING_VALUE;
 
 /**
  * Configuration for a LatLngString field.
@@ -657,7 +661,7 @@ export function firestoreLatLngString(config?: FirestoreLatLngStringConfig) {
   const transform = latLngStringFunction({ precision, wrap: false, validate: true });
 
   return firestoreString<LatLngString>({
-    default: defaultValue || DEFAULT_FIRESTORE_LAT_LNG_STRING_VALUE,
+    default: defaultValue || DEFAULT_LAT_LNG_STRING_VALUE,
     defaultBeforeSave,
     transform
   });
@@ -800,3 +804,12 @@ export function firestoreUnitedStatesAddress() {
     toData: firestoreUnitedStatesAddressAssignFn
   });
 }
+
+// MARK: Zoom
+export const MIN_FIRESTORE_MAP_ZOOM_LEVEL: ZoomLevel = 0;
+export const MAX_FIRESTORE_MAP_ZOOM_LEVEL: ZoomLevel = 22;
+
+/**
+ * Convenience function for firestoreNumber() for storing an integer ZoomLevel value.
+ */
+export const firestoreMapZoomLevel = firestoreNumber<ZoomLevel>({ default: 5, transform: { precision: 1, bounds: { min: MIN_FIRESTORE_MAP_ZOOM_LEVEL, max: MAX_FIRESTORE_MAP_ZOOM_LEVEL } } });